cambiata mappa, aggiunti getTarget bug("si incastra nello spawn se con cordinate 9, 11")

+++ src/entity/PacDot.java
@@ -23,7 +23,7 @@ public class PacDot extends Entity
     }
     public void setDefaultValues()
     {
-        dotSize = 2 * gp.scale;
+        dotSize = (int)((double)2 * gp.scale);
         alive = true;
         points = 10;
